requirements.txt
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
twilio>=9.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pytz>=2024.1
